Rename folder name (#3367)
