+++ src/blocks/social-share-child/deprecated.js
@@ -7,7 +7,7 @@ import classnames from 'classnames';
 import attributes from './attributes';
 import renderSVG from '@Controls/deprecatedRenderIcon';
 import links from './deprecatedLinks';
-import V2_6_9Save from './deprecated/V2_6_9/save';
+import V2_7_1Save from './deprecated/V2_7_1/save';
 
 const deprecated = [
 	{
@@ -51,7 +51,7 @@ const deprecated = [
 	},
 	{
 		attributes,
-		save: V2_6_9Save,
+		save: V2_7_1Save,
 	}
 ];
